refactor: ExcalidrawAnimatorEditor into 5 files
